+++ libs/flexIPC/fleXdIPCCommon/ChangeLog.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## [0.1.0] - 2018-07-26
+### Added
+- New virtual functions isServer() and isClient() for FleXdIPC class [Martin Strenger]
+
+### Changed
+- None
+
+### Removed
+- None
+
 ## [0.1.0] - 2018-07-17
 ### Added
 - None